Turret Update
Turrets now lock on to zombies in area using sphere-trigger-colliders. Those with the tag "Zombie" then are added to a list. using a while loop, we check if the zombie doesn't exists. if true, we lock onto the zombie  and kill it. once killed, remove the empty element to prevent Null Exception issues.